upd with some correctings
