+++ src/views/Home/HomeTrainings/HomeTrainings.jsx
@@ -7,32 +7,10 @@ import {
   OffsetImageBorder,
   ModalButton,
 } from '@/components';
-import homeTrainingsData from 'src/data/homeTrainingsData.json';
-import Trainitem1 from 'public/icons/home-tren-item-1.svg';
-import Trainitem2 from 'public/icons/home-tren-item-2.svg';
-import Trainitem3 from 'public/icons/home-tren-item-3.svg';
-import Trainitem4 from 'public/icons/home-tren-item-4.svg';
-import Trainitem5 from 'public/icons/home-tren-item-5.svg';
+import { TrainingsList } from './TrainingsList';
 import * as st from './HomeTrainings.module.css';
 
 export const HomeTrainings = ({ trainings }) => {
-  const Trainitems = {
-    1: <Trainitem1 />,
-    2: <Trainitem2 />,
-    3: <Trainitem3 />,
-    4: <Trainitem4 />,
-    5: <Trainitem5 />,
-  };
-  const trainingsList = homeTrainingsData?.map(el => {
-    return (
-      <>
-        <li key={el.id.toString()} className={st.item}>
-          <span className={st.itemImg}>{Trainitems[el.id]}</span>
-          <span className={st.itemText}>{el.text}</span>
-        </li>
-      </>
-    );
-  });
   return (
     <section className={st.section}>
       <Container>
@@ -41,7 +19,7 @@ export const HomeTrainings = ({ trainings }) => {
           <ReactMarkdown className={st.variant}>
             {trainings.variantFirst}
           </ReactMarkdown>
-          <ul className={st.listTrainings}>{trainingsList}</ul>
+          <TrainingsList />
           <div className={st.contentWrapper}>
             <OffsetImageBorder className={st.imageBorder}>
               <Image
